+++ demo/index.blade.php
@@ -1,12 +1,19 @@
 {{--test--}}
-
+{{--
+    test
+    --}}
 {{'one'}}
+{{
+'one'
+    }}
 {!!'one'!!}
+{!!
+'one'
+    !!}
 
 @if(true)
     hello
 @endif
-
 @php
     $one = 'test';
     $two = [
@@ -17,8 +24,13 @@
     $thr = $one||$two;
     $four = $one&&$two;
     $five = $one?:$two;
-    $five = $one??$two;
+    $six = $one??$two;
+    $svn = $one?'yes':$two;
+    $svn = $one ? 'yes' : $two;
+    $svn = $one
+    ?'yes'
+    :$two;
 @endphp
 <hr>
 
-@include('',['var'=>['one','two']])
+@include('',['var'=>['one',1,$two]])