+++ CHANGELOG.md
@@ -3,3 +3,9 @@
 ## 0.0.1
 
 - Initial release
+
+## 0.0.2
+
+- better regex for items on new line
+- add config & fix name space
+- update demo file